requirements.txt
Package          Version
---------------- -------
bcrypt           3.2.0
cffi             1.15.0
click            8.0.3
colorama         0.4.4
Flask            2.0.2
Flask-SQLAlchemy 2.5.1
greenlet         1.1.2
itsdangerous     2.0.1
Jinja2           3.0.2
MarkupSafe       2.0.1
numpy            1.21.3
pandas           1.3.4
pip              21.3.1
pycparser        2.21
PyMySQL          1.0.2
python-dateutil  2.8.2
pytz             2021.3
setuptools       49.2.1
six              1.16.0
SQLAlchemy       1.4.26
Werkzeug         2.0.2
